requirements.txt
asgiref==3.2.10
requests==2.26.0
django==2.2.19
django-cors-headers==3.14.0
django-filter==2.4.0
djangorestframework==3.12.4
djoser==2.1.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-import-export==2.7.1
djangorestframework-simplejwt==5.2.2
gunicorn==20.0.4
Pillow==9.5.0
psycopg2-binary==2.8.6
python-dotenv==0.19.2
pytz==2020.1
sqlparse==0.3.1 

